Refactor indexer to trigger subsystem and update related components
