+++ src/main/java/frc/robot/subsystems/trigger/TriggerIO.java
@@ -1,14 +1,14 @@
-package frc.robot.subsystems.indexer;
+package frc.robot.subsystems.trigger;
 
 import frc.lib.subsystem.BaseIO;
 import frc.lib.subsystem.BaseInputClass;
 import frc.lib.subsystem.IORefresher;
 import org.littletonrobotics.junction.AutoLog;
 
-public interface IndexerIO extends BaseIO<IndexerIO.IndexerIOInputs>, IORefresher {
+public interface TriggerIO extends BaseIO<TriggerIO.TriggerIOInputs>, IORefresher {
 
     @AutoLog
-    public static class IndexerIOInputs extends BaseInputClass {
+    public static class TriggerIOInputs extends BaseInputClass {
         public double motorRps = 0.0;
     }
